requirements.txt
asgiref==3.5.0
Django==4.0.3
psycopg2==2.9.3
sqlparse==0.4.2
djangorestframework==3.13.1
drf-writable-nested==0.6.3
